fix(checker): canonicalize DefId of barrel-re-exported generics in type-reference position
A generic interface or class that reaches a file through a barrel re-export
(`export { X } from` / `export type { X } from`) and is used in a
type-reference position (`const x: Base<number>`, a property annotation) was
rejected with a false `TS2322`: `{ value: number }` was not assignable to
`Base<number>`. A direct single-hop import was already correct.

Root cause: lowering a type reference builds `Application(Lazy(DefId), args)`.
The `DefId` must be keyed to the declaring `(SymbolId, file)`. When the name
reaches the file through a barrel, the name-based resolver derived the `DefId`'s
file from `decl_file_idx` / current-file-local heuristics and the
raw-`SymbolId`-keyed `symbol_to_def` cache (raw ids collide across binders), so
it attributed the `DefId` to the intermediate re-export file instead of the
declaration. That non-canonical `DefId` has no generic body registered, so the
application stayed opaque and its type parameter was never substituted.

Structural rule: when a bare type reference's local symbol is an import alias
reaching its declaration through a re-export chain, key the application base
`DefId` to the declaration's `(SymbolId, file)`. `reexported_declaration_def_id_for_lowering`
follows the chain (the same reference-aware resolver the heritage fix #13803
uses) and `def_id_for_declaration_in_file` mints/looks up the `DefId` keyed
strictly to the declaration, bypassing the cross-binder-colliding cache. Scoped
to interface/class declarations, whose bodies resolve on demand once the def is
canonical; re-exported generic type aliases need separate eager body
registration and are left on the existing path (no behavior change for them).

Advances the kysely (#10663) and valibot (#13212) false-positive families
(non-canonical cross-arena instantiation).

Two files at the 2000-line ceiling are split to make room:
`context/def_declaration.rs` (declaration-keyed DefId minting) and
`symbols/symbol_resolver_entity_name.rs` (text entity-name resolution).
