+++ crates/tsz-checker/Cargo.toml
@@ -33,6 +33,9 @@ dashmap = { workspace = true }
 web-time = { workspace = true }
 stacker = "0.1.23"
 [[test]]
+name = "reexport_generic_typeref_instantiation_tests"
+path = "tests/reexport_generic_typeref_instantiation_tests.rs"
+[[test]]
 name = "cross_file_recursive_alias_intersection_tests"
 path = "tests/cross_file_recursive_alias_intersection_tests.rs"
 [[test]]
