+++ crates/tsz-checker/src/context/mod.rs
@@ -35,6 +35,7 @@ pub(crate) mod env_eval_cache;
 mod file_session_reset;
 pub mod lifetime_shells;
 pub use lifetime_shells::{FileSession, LspPersistentCache, SpeculationScope, WorkerContext};
+mod def_declaration;
 mod def_mapping;
 mod def_mapping_formatters;
 mod deferred_flow_env_write;
