+++ crates/tsz-checker/src/symbols/mod.rs
@@ -3,5 +3,6 @@ pub(crate) mod name_text;
 pub mod scope_finder;
 pub(crate) mod scope_finder_contexts;
 pub mod symbol_resolver;
+pub(crate) mod symbol_resolver_entity_name;
 pub(crate) mod symbol_resolver_qualified;
 pub(crate) mod symbol_resolver_utils;
